尝试解决 Issue#8 时留下的一些中间产物。对某些类型实现了 IDispose，手动编写了 ContractSerializer 。这些内容会在下一个提交中被清理。 合并了 BOP 的几个测试样例。 Most of all，尝试对 MagikConsole 启用服务器GC。

+++ MAGIK/Academic/Contract/Entity.cs
@@ -10,7 +10,7 @@ namespace Microsoft.Contests.Bop.Participants.Magik.Academic.Contract
     public class Entity
     {
         [JsonProperty("logprob")]
-        public double LogarithmProbability { get; set; }
+        public float LogarithmProbability { get; set; }
 
         [JsonProperty("Id")]
         public long Id { get; set; }
